+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## unreleased
 
+* Local Payments
+  * Add `displayName` to `LocalPaymentRequest`
+
 * Venmo
   * Make `VenmoRequest` parcelable
 * ThreeDSecure